Add files via upload
Boardo is an IOT device which is used to count the no of users who read
a particular notice from a notice board. In making of this project I came across few
difficulties using sensors like ultrasonic which later came to know that was not right
one and shifted to long range IR which gave required results up to better extents. For
boardo I used Arduino as the main component and coded in embedded c.